refactor spacy_parse
