+++ project/urls.py
@@ -20,7 +20,7 @@ from django.conf.urls.static import static
 from django.urls import path
 from django.urls.conf import include
 from users.views import index
-from .views import health, upload, parse
+from .views import health, upload
 
 urlpatterns = [
     path('admin/', admin.site.urls),
@@ -28,10 +28,10 @@ urlpatterns = [
     path('accounts/', include('allauth.urls')),
     path('quality_control/', include('quality_control.urls')),
     path('masterdata/', include('masterdata.urls')),
+    path('parse/', include('spacy_parse.urls')),
     path('', index, name='index'),
     url(r'^health$', health),
     path('upload/', upload, name='upload'),
-    path('parse/', parse, name='parse'),
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
 
 if settings.DEBUG:
